dynamic-load: move registrar to sdk dir
